moved scripts into audit module, added audit helper and generic auditor
