+++ audit/delete_sequence_files_from_audit.py
